[ao][sparsity] comsability for sparsity and QAT convert
Summary: The primary issue for enabling sparsity to work with QAT
convert (unlike normal quantization convert) is that when the
parametrized module undergoes the QAT convert, the parametrizations need
to be maintained. If the parametrizations don't
get transfered during the convert, the sparsifier would lose its
connection to the model. In practice this was handled using the
transfer_parametrizations_and_params function to identify all
parametrizations on the original module and then move them (and their
associated parameters) to the new module.

Test Plan: python test/test_ao_sparsity.py TestComposability

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 2befc87f42720e8603b1640b0ccfd9813a453e9e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74848